+++ Changelog.md
@@ -1,8 +1,18 @@
 # Change Log
 
+## 1.0.1 - _19.06.2026_
+
+**Patch Release**
+
+- Some bug fixes
+- Device removal on close/disconnect
+- Improved reconnect logic
+- WebSocket cleanup improvements
+- Web Browser API improvements
+
 ## 1.0.0 - _03.11.2025_
 
-**Release**
+**Major Release**
 
 - Added a popup to set port other than 3210
 - Auto-reconnect on port changed (via save button)